Restructure command parsing to use multiple command classes
Command parsing is done in a while loop.

This locks the thread which waits for user input and is not
compatible with the gui interface.

Extracting each commmand into a class allows us to model the chat
bot without the use of a for loop. It clarifies what input is
expected at each state of the bot.