Enable FileNotFoundException message shown in GUI
When storage files to be read by the application to load the tasks are
not found, an exception is thrown and the error message is not shown in
the GUI by the chatbot, but shown in the terminal.

This is poor user feedback, undermining user experience as app users
will not know that a new file is created rather than loading a
supposedly existing file for saving or loading tasks.

Let's add the exception message into the dialog Container text parameter,
allowing the message to be shown in the GUI. To do that:
* enable the chatbot to have a welcome Message attribute which changes
  if there is an exception
* modify code in Main.java and MainWindow.java such that the UI element
  that indicates the exception message is only rendered after the
  instance of the bot is addded to the MainWindow.

This allows the Exception message to be shown in the GUI to the users.