Add streams to modified files
No use of streams in current files.

Using streams for parts of the application can make
it easier for us to reason about and debug the program.

Let's change parts that is changable to streams into
streams.

This will make those parts less error prone and easier
to reason about.