Add markMultipleTasks method to TaskList class to return processed tasks and add Javadoc
There is no method to mark multiple tasks at once in the TaskList class.

Allowing the user to mark multiple tasks be useful as the user does not need to key mark so many times is he wants to mark or unmark multiple tasks.

Create the markMultipleTasks method to return an ArrayList<Task> with the tasks that are marked or unmarked. Add Javadoc comments to describe the method’s parameters, return value, and exceptions.

Returning the processed tasks allows for better tracking of which tasks were affected which can be used by the UI class to display the information.

Error handling for NumberFormatException and IndexOutOfBoundsException is included, throwing InvalidInputException and InvalidTaskNumberException respectively. This ensures robust handling of invalid input and out-of-bounds indices.