Add assertions to TaskList and Ui classes
In the Ui class, an assertion is made in the processDescription method
that the current command is an add command, since the only command
that will lead to the method being called is the AddCommand command.

In the TaskList class, assertions are made in some of the methods, that
indexes provided are not out of bounds, since whether the index is
out of bounds is already checked by the Ui class.