Improve code quality
The current code quality is decent but there is room for improvement.

There are some complex methods and there is room to make the happy
path more apparent in the code.

Lets reduce the nesting in these complicated methods specifically
in the Processor.java to make the happy path more apparent.

This makes the code alot more readable and less confusing for future
developers.