Update files to improve code quality
The execute methods for most of the Command classes were too long.

Extracting the AddCommand Class to AddToDoCommand Class,
AddDeadlineCommand Class and AddEventCommand Class removes the need
to check for whether the AddCommand is for a To-Do task, Deadline
task or a Event task. Thus shortening the execute method for these
classes.

Most of the Command classes had their execute method shortened by
splitting the code body into several methods. Some of these
methods are created in the class itself while some were created in
the Parser Class instead since it was a common method for the Command
classes.