Code Refactoring
There is a lot of accumulated technical debt as until now, the focus
was on getting the code to work properly.

This made some of the code be redundant and other parts of the code
to be very long.

Let's refactor the code and re-organise the methods and classes to
ensure that the code organisation and quality improves.

This will reduce the technical debt accumulated, making the codebase
more useable for the future.