Add Pair class
It is not possible to return two items of different types in a function
in a pair easily.

Some functions in Storage class need to return an Optional either
containing something or nothing, and an accompanying String message
which may be needed for Knight2103 as an error message.

Let's Add a Pair class which represents like a type that contains 2
different types.

So that it is possible to return 2 items of different types
simultaneously from a single function.