Storage class: Change return type to Pair
The bot is not be able to show all error messages to the users in the
GUI when the storage file is loaded, especially for errors that arise
due to incorrect formatting of contents in the storage file.

It is important that a comprehensive user feedback is given in the GUI,
thus code change is needed to enable that. While throwing an exception
works as well, it is not done that way because once an exception is
thrown, the entire code just stops -- i.e. the loading of file contents
will end prematurely where the next few lines in the file which are in
the correct format are not processed to create the list as comprehensive
as possible. We want to process every line in the file regardless if the
line with wrong formatting is in the middle of the correct lines.

Let's change the return type of the function such that errors in the
file contents formatting are not exactly shown as exceptions, but
saved in the bot's attribute called error message.

These error messages are created in the functions of the Storage class
which will be eventually be passed into the bot's error message
attribute, enabling the bot to show the exception message when the app
is first started.