Add flexible inputs
The commands are very rigid and not customisable.

This increases user frustration when they misspell a single word in
a long command.

Let's add functionality to make the commands more flexible and be
called on using different words.

This makes the user experience better and more customised to each
user.