Improved First Startup Experience
The necessary storage files were not being created on first start of
the app.

The user will be confused at the error message and not want to use
the app anymore.

Let's improve the first startup experience and enable all necessary
file creations upon the first creation of the app.

This will make the initial user experience much better and smoother
for new users.