Remove ArrayIndexOutOfBounds Exception in catch
Catching of ArrayIndexOutOfBounds Exception in ModifyCommand class is to
catch for any errors where the command are not segmented in less than 2
parts. However, such errors should have been surfaced in the Parser.java
file and to be caught at that point in time.

Thus, there is no need to duplicate code.

Let's remove the catch statement for ArrayIndexOutOfBounds Exception in
ModifyCommand.java

This is to prevent code duplication.