Detect blank lines in Storage File not as errors
If the Storage textfile contains blank lines in between the list of
tasks written in the textfile, the code will detect these blank lines
as errors as they do not start with "T", "D", or "E".

This will confuse users as user feedback will say there is a need to
change the storage file contents formatting due to an error, but such
blank lines are minute deviations from the ideal formatting and may
frustrate users the limited flexibility of the formatting required in
the text file.

Let's remove such a restriction by allowing the code to check if it is a
blank line. If so, the code will automatically read the next line in the
file without further checking.

This is to allow users to be able to leave blank lines in the storage
text file, enabling flexibility in the file contents formatting,
enhancing user experience.