Update Exception messages shown in GUI
Messages written by the bot is not very user friendly as users do not
exactly know what went wrong. It will be better if these messages
indicate clearly for users to understand how they can input the correct
commands.

Let's edit the exception messages with more clarity.

This is to enhance user experience.