Refactor code via inline method
checkForMarkedFormat() is a helper function to check if the task to be
added should be marked or unmarked. If marked, the task to be added
need to have its isDone status to be updated before the task is added
into the list of tasks.

For code readability, SLAP principle is to be adhered and it seems that
such code violates SLAP principle. Since this helper function code
manages the interpretation of each line, of the same level of
abstraction as that in convertLineToTask(). Thus, this code will not be
extracted as a helper function but directly placed in
convertLineToTask().

Let's shift the code in the helper function checkForMarkedFormat()
into convertLineToTask() function.

This is to adhere to SLAP principle as much as possible.