Rename functions and class names for readability
Certain functions are not very clear in conveying the meaning or their
purpose. Classes as well. For example, the fact that Todo class is
inherited from Task class can be made clear by having the Todo class
being named as TodoTask. The word "print" in a function name may
suggest that the function return type is void as the statement is to be
printed only in the console. But for some functions with this "print"
word, it was meant to return a string.

Let's change the namining to convey more clearly about the
functionality and relationships of these functions and classes.

This is to improve code quality.