Move functions that stringify TaskList elsewhere
Functions formatToFullList(), formatToMatchedList(), formatToFileList()
are functions that make the TaskList into String format depending on
whether the list of tasks to be written is in the File Storage, or in
the GUI. And if it is in the GUI, if it is the full list of task or
selected tasks only. These functions are dependent on the context
taskList is to be shown, thus these functions may not be the best to be
in the TaskList class, and to be shifted to the Ui class and Storage
class -- the class that is relevant.

For formatToFullList() and formatToMatchedList(), both have similar code
logic which can be extracted out into a helper function, adhering to
DRY princniple.

Let's
* extract the common patterns of formatToFullList() and
  formatToMatchedList()
* shift the code to more relevant classes

This is to adhere to DRY principle and OOP design practices better.
Hopefully.