Add assertions to check that filepaths are valid
There is no way to automatically check if filepaths
(especially resource files) provided in Constants.java
are valid or existent.

Therefore, validating these files need to be done
manually, otherwise the validity of these filepaths
are left unknown until runtime.

Thus, provide a function that automates these checks
(implemented in the main function of Constants.java).

As this is the only validation done by Constants.java,
it is adequate and convenient to execute the check
through the main function of Constants.java.