Add assertions to Storage, Deadline and Event files
CSV file for data storage is assumed to already have the proper headers, while the deadlines and relevant time periods entered by the user are assumed to be logical timings for both the deadline and event tasks created.

Inappropriate dates entered may render the tasks created useless and of little utility to the user, and may perhaps interfere with future iterations to the project which leverage on the created tasks having proper and logical dates.

Assertions have been added to ensure only deadlines past the time the user creates the task can be used. Same goes with the duration of event tasks as indicated by the user.