Remove magic numbers or literals
Some code use certain numbers or literals which its meaning may not
be the most obvious. To make the code more readable and understandable,
these numbers or literals are stored in constant variables named in
such a way that represents the meaning.

Let's replace these magic numbers or literals with constants.

This is to make the code more easy to understand.