Refactor loops to stream for list command
The current implementation of the list command to generate the display
string of the task list uses a for loop.

The for loop implementation requires the explicit use of a StringJoiner
and requires explicit management of how the strings are joined.

Refactor the loop to use java streams instead.

The resultant code that uses the stream is simplier (in my opinion),
without the explicit use of StringJoiner and explict adding of strings
to the StringJoiner object. Rather, an alternate approach to declare
that these stream of strings are to be joined together via the
`Stream.collect` and `Collectors.joining` methods.