Add assertions to classes
Some assumptions in the code are only verified if a bug occurs.

This can cause one to be less certain of the robustness of his code,
even making debugging harder as he has to go through each assumption
and verify that they are not the cause of the bug.

Using assertions in the classes allows us to check if a certain
behaviors are valid at each point in the code.

By doing so, one can have more confidence in the code and narrow down
the cause of a new bug to areas where assertions have failed or
areas areas not sufficiently covered by assertions, reducing areas
in the code in which one has to search for the bug.