Add assert function into the Storage object to check that there is a actual file in the Storage
This is to follow the instruction of the course and make the code safer

Add an assertion line when creating new instance of Storage item

The code should not continue further if there is no memory file and fail to create.
The assertion is to check that the code is continue with a memory file.