Improve code quality
Code has certain flaws, with missing comments and lack of
prioritization of the happy path being examples of them.

Poor code quality makes it difficult to find bugs and correct them.
Lack of comments, for example, maintains the obscurity of the code,
increasing the difficulty of understanding why the code was written in
a certain way.

Make the happy path more prominent and describe the ErrorBob and
InvalidCommandException classes, along with the parseDatetime method.

Making the happy path more prominent reduces nesting while comments
on the specified classes and method reduce ambiguity about the
roles they perform.