Add shorter aliases for command keywords
Users have to type out the full command keyword (Eg. unmark 2), which
can be quite troublesome if the user uses the chatbot a lot.

By allowing for shorter aliases for command keywords (Eg u instead
of unmark), experienced users will be able to type lesser to do
what they want to do with the chatbot.