Clean up storage for future debug on list command
The list command returns "Here is your list" even when the list is
empty.

The behaviour is incorrect, as no list should be displayed when it is
empty.

Clean up storage to prepare for debugging and  fixing this issue.

Proper storage cleanup is needed to accurately reflect list data and
resolve the false output.

Future debugging will focus on identifying the source of the incorrect
list ouput.