Abstract Yapper commands and add bind and unbind commands to help users add custom commands.
There are several commands for the Yapper chatbot and putting them all into one method with a big if-else block becomes messy and unreadable.

We should try to abstract and turn each command into it's own class and package them separately so code becomes more readable and easier to manage.

As a step towards making the code more readable, let's turn all the commands into classes so that binding and unbinding new commands can also be made easier.