Add Assertions
No assertions present in the code.

Invariants in the program cannot be tested before chatbot is run,
meaning higher chance of unchecked exceptions.

Let's use Java's assert to check for invariants in the Storage
class like whether the save file is created after saving or if it
exists before loading from it.