Add Stream to read secret file
No streams present in the current program.

Without the versatility of streams, errors and null values cannot
be handled efficiently and lazily.

Let's add streams when reading the secret file in the Jackson.java
file.

This is done to leverage the Optional class in Java for lazy
loading, in case the secret file text changes and becomes long and
more inefficient to load.