Sort: Integrate sort functionality
Users were not able to sort tasks according to the task attributes.

This lack of organisation can make it hard to organise tasks in
a neat way according to user preference, which defeats the point of
the chatbot.

Let's,
* create a new sort method in TaskList
* make CustomDateTime comparable to facilitate comparison
* add new sort command to Parser, Ui and Jackson
* add Javadocs
* do minor spring cleaning of code comments and existing Javadocs

Directly incorporating comparison functionality into the classes
involved is much better than simple text comparison of the classes
as it is more airtight and reliable.