Rebuild jar file after accounting for checkstyle
Checkstyle picks up import order and spacing errors in some files. It also notes that some methods are missing JavaDoc comments.

Additionally, the checkstyle files are not picked up by gradle.

This means that some parts of the code have some styling errors, and gradle is unable to check for them when building the jar file.

Let's edit the code in such a way that the checkstyle files are picked up by gradle and styling tests are run upon building the jar file.
Also add JavaDoc comments to more methods and fix the import order and spacing issues present in some files.

This allows the code to adhere to the coding convention specified and prevents the build from failing.

Refer to [this link](https://se-education.org/guides/conventions/java/index.html) for the Java coding standard.