Add assertions to TaskReader.java
The TaskReader will read the information from the file to determine the
tasks to be added. It is expected to throw a TypeException when it reads
an invalid type from the file. If there was an invalid type, but no
TypeException is thrown, an error will occur.

Let's assert such that upon reading the information from the file, the task
type is valid before proceeding.