Add find duplicate logic to Jar bot
Let's create a method called findDuplicate in TaskList, which will check if the task entered
exists in the list of tasks. If so, we will return a message to user indicating that they cannot add
the task.

Let's also add some getter and setter methods in the task classes to extract information from the tasks
while adhering to OOP conventions.