Use assert feature and edit javadocs comments
Assertions are currently not used.

Assert feature document important assumptions that should hold at
various points in the code. Assertions are suitable for verifying
assumptions about Internal Invariants, Control-Flow Invariants,
Preconditions, Postconditions, and Class Invariants. An assertion
failure indicates the programmer made a mistake in the code.

Assertions are now added to Ui.java only as most of the checking
needs to be done in this class.

JavaDocs comments were edited as many return types of methods were
changed in Ui.java