Refactor Parser class
Parser has over 300 lines of code. The logic behind Parser is
solely coded in Parser.java. This could lead to difficulty in
debugging as there are a lot of lines to read.

Extract the logic of Parser class using a new Function package.

Depending on the command given by user, Parser will instantiate
a suitable Function to process the arguments of the user.
Depending on what was typed by the user as arguments, the Function
will return a suitable Executable.