The methods extractTimeDetail and extractDescription has for loops and conditional statements to get the respective details.
The current implementation is harder to read and it is not as flexible
as compared to Java streams.

Change the two methods to use Java Streams for processing and getting
the necessary details.

Streams provide a variety of options and it makes the code more
readable and easier to maintain.