Improve code quality
Methods and variable names are not representative of what the actual method and variables do
which may lead to bug inducing code due to misunderstanding of what the variable actually do.

Method names and variable names are renamed to better reflect their functionality and purpose.
Group static variable together with common prefixes

Changes was done to ensure anyone working on the project is able to understand what is written with little to no ambiguity