Enhance Code Quality
The main bulk of arrowhead code in the Parser parseCommand
method was enhanced to using the SLAP principle. Several
private static methods were extracted for this purpose.

A bug was also fixed in the getTaskIndex method where user not
inputting an index was not accounted for. The new guard
statement will raise an InvalidCommandException when this
specific user error is made.