Add feature to detect duplicate tasks when adding tasks
Currently, the bot is unable to check whether the user has added
a duplicate task, which could cause a hassle for the user to
manage their tasks.

It is important that we detect and prevent duplicates from existing
in the task list to prevent unintended actions.

To solve this, Task and its subclasses will now have a
'checkDuplicate' function to check if the tasks are similar.
When the user enters a duplicate task, the command will initiate
a check against all tasks in the task list to check if there are
any duplicates before allowing the user to add the task.
If duplicates are detected, the bot will inform the user.

This solution ensures that only unique tasks will be stored in the
task list for users.