Assign error messages to constants where possible
Most of the error messages are passed to the exception constructor
directly.

Extracting these message into constants makes the code more readable as
error messages that pertain to that specific exception.

This removes the magic literal strings, and makes it easier to
understand at a glance.

However, not all error messages could be extracted into constants so
easily, especially those that use method parameters in their lines.

This applies to Parser.java at this point of time.